graphql: genny was generating code with wrong import, removed genny
