+++ tools.go
@@ -4,7 +4,6 @@ package tools
 
 import (
 	_ "github.com/99designs/gqlgen"
-	_ "github.com/cheekybits/genny"
 	_ "github.com/praetorian-inc/gokart"
 	_ "golang.org/x/vuln/cmd/govulncheck"
 )
